Enable the check model command when model pane is open
When the model pane is active, the last checked model is used as the model
to check. This is needed since the spec to check is otherwise retrieved
using the active file. Since there is no active file if the model pane
is active, we need to supply another file to check.

This commit also refactors several checkModel helper functions to return
Result types. This allows for errors to be propagated back and
potentially ignored rather than immediately resulting in an error
message.